chore: 거래 비용 계산식 변경

+++ back/src/main/java/com/rezero/rotto/entity/TradeHistory.java
@@ -49,5 +49,5 @@ public class TradeHistory {
 	private String bcAddress;
 
 	@Column(name = "token_price")
-	private int tokenPrice;
+	private double tokenPrice;
 }
